解决sonarlint问题

+++ src/main/java/com/stssm/github/io/javaweb/aaa024javaweb20220723/Bbb006CheckCodeUtil.java
@@ -211,10 +211,10 @@ public class Bbb006CheckCodeUtil {
 		int phase = RANDOM.nextInt(2);
 
 		for (int i = 0; i < h1; i++) {
-			double d = (double) (period >> 1)
+			double d = (period >> 1)
 					           * Math.sin((double) i / (double) period
-							                      + (6.2831853071795862D * (double) phase)
-									                        / (double) frames);
+							                      + (6.2831853071795862D *  phase)
+									                        /  frames);
 			g.copyArea(0, i, w1, 1, (int) d, 0);
 			if (borderGap) {
 				g.setColor(color);
@@ -233,10 +233,10 @@ public class Bbb006CheckCodeUtil {
 		int frames = 20;
 		int phase = 7;
 		for (int i = 0; i < w1; i++) {
-			double d = (double) (period >> 1)
-					           * Math.sin((double) i / (double) period
-							                      + (6.2831853071795862D * (double) phase)
-									                        / (double) frames);
+			double d =  (period >> 1)
+					           * Math.sin( i /  period
+							                      + (6.2831853071795862D *  phase)
+									                        /  frames);
 			g.copyArea(i, 0, 1, h1, 0, (int) d);
 			if (borderGap) {
 				g.setColor(color);
